+++ org.lxtk.lx4e/src/org/lxtk/lx4e/model/impl/LanguageSourceFile.java
@@ -48,7 +48,6 @@ import org.eclipse.handly.util.Property;
 import org.eclipse.jface.text.Document;
 import org.eclipse.lsp4j.DocumentSymbol;
 import org.eclipse.lsp4j.DocumentSymbolParams;
-import org.eclipse.lsp4j.SymbolInformation;
 import org.eclipse.lsp4j.SymbolKind;
 import org.eclipse.lsp4j.jsonrpc.messages.Either;
 import org.lxtk.DefaultWorkDoneProgress;
@@ -432,6 +431,7 @@ public abstract class LanguageSourceFile
         return null;
     }
 
+    @SuppressWarnings("deprecation")
     private List<DocumentSymbol> getDocumentSymbols(DocumentSymbolProvider provider,
         URI documentUri, IProgressMonitor monitor) throws CoreException
     {
@@ -444,7 +444,7 @@ public abstract class LanguageSourceFile
         request.setUpWorkDoneProgress(
             () -> new DefaultWorkDoneProgress(Either.forLeft(UUID.randomUUID().toString())));
 
-        List<Either<SymbolInformation, DocumentSymbol>> result;
+        List<Either<org.eclipse.lsp4j.SymbolInformation, DocumentSymbol>> result;
         try
         {
             result = request.sendAndReceive();
@@ -459,7 +459,7 @@ public abstract class LanguageSourceFile
             return Collections.emptyList();
 
         List<DocumentSymbol> symbols = new ArrayList<>(result.size());
-        for (Either<SymbolInformation, DocumentSymbol> item : result)
+        for (Either<org.eclipse.lsp4j.SymbolInformation, DocumentSymbol> item : result)
         {
             if (item.isRight())
                 symbols.add(item.getRight());