+++ org.lxtk.lx4e/src/org/lxtk/lx4e/requests/DocumentSymbolRequest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020, 2021 1C-Soft LLC.
+ * Copyright (c) 2020, 2022 1C-Soft LLC.
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which is available at
@@ -25,6 +25,7 @@ import org.lxtk.DocumentSymbolProvider;
 /**
  * Requests information about symbols defined in the given text document.
  */
+@SuppressWarnings("deprecation")
 public class DocumentSymbolRequest
     extends LanguageFeatureRequestWithWorkDoneAndPartialResultProgress<DocumentSymbolProvider,
         DocumentSymbolParams, List<Either<SymbolInformation, DocumentSymbol>>>