+++ org.lxtk/src/org/lxtk/WorkspaceSymbolProvider.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020, 2021 1C-Soft LLC.
+ * Copyright (c) 2020, 2022 1C-Soft LLC.
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which is available at
@@ -15,9 +15,10 @@ package org.lxtk;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
-import org.eclipse.lsp4j.SymbolInformation;
+import org.eclipse.lsp4j.WorkspaceSymbol;
 import org.eclipse.lsp4j.WorkspaceSymbolParams;
 import org.eclipse.lsp4j.WorkspaceSymbolRegistrationOptions;
+import org.eclipse.lsp4j.jsonrpc.messages.Either;
 
 /**
  * Provides information about project-wide symbols matching a query string.
@@ -33,8 +34,9 @@ public interface WorkspaceSymbolProvider
      * @param params not <code>null</code>
      * @return result future (never <code>null</code>)
      */
-    CompletableFuture<List<? extends SymbolInformation>> getWorkspaceSymbols(
-        WorkspaceSymbolParams params);
+    @SuppressWarnings("deprecation")
+    CompletableFuture<Either<List<? extends org.eclipse.lsp4j.SymbolInformation>,
+        List<? extends WorkspaceSymbol>>> getWorkspaceSymbols(WorkspaceSymbolParams params);
 
     /**
      * Returns the context object associated with this provider.